Test Case for Radiation At Centre A Circle
The test case solves for the radiation generated at the centre of a circle. the circle is the trajectory of the particle. This case is completely analytically solvable. The radiation magnetic field should reduce to the case of a current carrying loop if the electron is assumed as a source of current.